% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma_collapse.R
\name{comma_collapse}
\alias{comma_collapse}
\title{Collapses input (as rows) from clipboard with comma sign and adds square brackets.}
\usage{
comma_collapse(brackets = TRUE)
}
\arguments{
\item{brackets}{TRUE by default. If TRUE, inserts square brackets around each row. If FALSE, does not insert brackets.}
}
\value{
One cell with rows collapsed by comma sign.
}
\description{
This function gets rows from the clipboard and outputs the rows (with square brackets) connected by plus signs.
}
\examples{
\dontrun{
#Run function without square brackets
comma_collapse(brackets = FALSE)
}
}
\author{
Michael Stoller <mstoller84@gmail.com> <michael.stoller@ctu.unibe.ch>
}
