% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.table.R
\name{copy.table}
\alias{copy.table}
\title{Get table output to clipboard for easy pasting into excel.}
\usage{
copy.table(obj, size = 4096, col.names = FALSE)
}
\arguments{
\item{obj}{}

\item{size}{}
}
\value{
A table is returned.
}
\description{
This function gives a table output to the clipboard.
}
\author{
NA
}
