% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_choices.R
\name{dict_choices}
\alias{dict_choices}
\title{Creates choices from rows.}
\usage{
dict_choices(stepsize = 1, stepmin = 1)
}
\arguments{
\item{stepsize}{1 by default. The stepsize by which to increment the choice value from one choice to the next.}

\item{stepmin}{1 by default. The value for the first choice}
}
\value{
A one-cell string with formatting according to choices in REDCap dictionary.
}
\description{
This function gets variables (in rows) from the clipboard and outputs choices for pasting into the REDCap dictionary.
}
\examples{
\dontrun{
#Run function with stepsize 1 and initial step 0
dict_choices(stepsize = 1, stepmin = 0)
}
}
\author{
Michael Stoller <mstoller84@gmail.com> <michael.stoller@ctu.unibe.ch>
}
