% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus_collapse.R
\name{plus_collapse}
\alias{plus_collapse}
\title{Collapses input (as rows) from clipboard with plus sign and adds square brackets.}
\usage{
plus_collapse()
}
\arguments{
\item{brackets}{TRUE by default. If TRUE, inserts square brackets around each row. If FALSE, does not insert brackets.}
}
\value{
One cell with rows collapsed by plus sign.
}
\description{
This function gets rows from the clipboard and outputs the rows (with square brackets) connected by plus signs.
}
\examples{
\dontrun{
#Run function without square brackets
plus_collapse(brackets = FALSE)
}
}
\author{
Michael Stoller <mstoller84@gmail.com> <michael.stoller@ctu.unibe.ch>
}
