% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_lines.R
\name{repeat_lines}
\alias{repeat_lines}
\title{Create repetitions for redcap dictionary}
\usage{
repeat_lines(repeatn = 1)
}
\arguments{
\item{repeatn}{Specify number of repetitions. 1 by default (ouput = input)}
}
\value{
Repetitions of the input
}
\description{
This function gets rows from the clipboard and outputs the repetitions back to the clipboard.
}
\examples{
\dontrun{
#Run function to create 3 repetitions
repeat_lines(repeatn = 3)
}
}
\author{
Michael Stoller <mstoller84@gmail.com> <michael.stoller@ctu.unibe.ch>
}
